workers/agent-worker/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
httpx==0.28.0
pydantic==2.10.0
anthropic==0.40.0
python-dotenv==1.0.0
